ovn-tester: Factor out modules from ovn_utils.py
